backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
psycopg2-binary==2.9.10
boto3==1.28.17
python-magic==0.4.27
gunicorn==21.2.0
Pillow==11.3.0